requirements.txt
et-xmlfile==1.1.0
ismrmrd==1.12.5
nibabel==5.1.0
numba==0.58.1
numpy==1.23.2
pandas==2.0.3
Pillow==10.0.0
pydicom==2.3.0
pyinstaller==6.3.0
pyMapVBVD==0.5.4
FreeSimpleGUI==5.2.0.post1
python-dateutil==2.8.2
python-git==2018.2.1
scikit-image==0.21.0
scipy==1.10.1
SimpleITK==2.3.1
tqdm==4.66.1

